some ui changes

+++ src/main/java/com/example/application/data/views/TaskView.java
@@ -1,21 +1,20 @@
 package com.example.application.data.views;
 import com.example.application.data.FormEvent;
+import com.example.application.data.RoleEnum;
 import com.example.application.data.entity.TaskEntity;
-import com.example.application.data.entity.UserEntity;
-import com.example.application.data.repository.TaskRepository;
 import com.example.application.data.service.TaskService;
 import com.example.application.data.service.UserService;
+import com.example.application.security.SecurityUtils;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.button.ButtonVariant;
-import com.vaadin.flow.component.checkbox.Checkbox;
 import com.vaadin.flow.component.grid.ColumnTextAlign;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.html.H1;
 import com.vaadin.flow.component.html.Header;
 import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.orderedlayout.FlexComponent;
+import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.textfield.TextField;
@@ -23,6 +22,7 @@ import com.vaadin.flow.data.value.ValueChangeMode;
 import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.Route;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
@@ -35,18 +35,22 @@ public class TaskView extends VerticalLayout {
 
     @Autowired
     TaskService taskService;
-
     @Autowired
     UserService userService;
+
 
     Grid<TaskEntity> grid = new Grid<>(TaskEntity.class);
     TextField filter = new TextField();
     TaskForm taskForm;
-
     public TaskView(TaskService taskService, UserService userService)  {
+        this.userService = userService;
+        this.taskService = taskService;
+        if (SecurityUtils.isAuthenticated()) {
+            Notification.show("Hello " + SecurityUtils.getName());
+        }
        H1 title = new H1 ("THIS SITE IS UNDER CONSTRUCTION");
        title.addClassName("title");
-        HorizontalLayout header;
+       HorizontalLayout header;
        /* if (SecurityUtils.isAuthenticated()){
             Button logout = new Button("Logout", click -> SecurityUtils.logout());
             header = new HorizontalLayout(logout);
@@ -54,13 +58,10 @@ public class TaskView extends VerticalLayout {
             header = new HorizontalLayout();
         }*/
        // add (header);
-
-        this.userService=userService;
-        this.taskService=taskService;
         setSizeFull();
         configureGrid();
         configureForm();
-        add(getToolBar(),getContent());
+        add(getToolBar(), getContent());
         updateFromFilter();
         closeEditor();
     }
@@ -87,15 +88,23 @@ public class TaskView extends VerticalLayout {
     }
 
     private void saveTask(FormEvent.SaveEvent event){
-        taskService.saveTask(event.getTask());
-        updateFromFilter();
-        closeEditor();
+        if (checkRoleLoggedIn()) {
+            taskService.saveTask(event.getTask());
+            updateFromFilter();
+            closeEditor();
+        } else {
+            Notification.show("You are trying to save a task that is not yours");
+        }
     }
 
     private void deleteTask(FormEvent.DeleteEvent event) {
-        taskService.deleteTask(event.getTask().getId());
-        updateFromFilter();
-        closeEditor();
+        if (checkRoleLoggedIn()) {
+            taskService.deleteTask(event.getTask().getId());
+            updateFromFilter();
+            closeEditor();
+        } else {
+            Notification.show("You are trying to delete a task that is not yours");
+        }
     }
 
     private void updateFromFilter() {
@@ -108,7 +117,8 @@ public class TaskView extends VerticalLayout {
         filter.setValueChangeMode(ValueChangeMode.LAZY);
         filter.addValueChangeListener(e -> updateFromFilter());
         filter.setWidth("2000px");
-        Button addTaskBtn = new Button("Add task",new Icon(VaadinIcon.PLUS));
+        Button addTaskBtn = new Button("Add task", new Icon(VaadinIcon.PLUS));
+
         addTaskBtn.addThemeVariants(ButtonVariant.LUMO_PRIMARY);
         addTaskBtn.addClickListener(click -> addTask());
         addTaskBtn.setWidth("200px");
@@ -125,10 +135,7 @@ public class TaskView extends VerticalLayout {
     }
 
     private void configureGrid() {
-        /*if(userService.getUsers() == null)*/
-
         grid.setSizeFull();
-
         grid.setColumns("title", "startDate", "deadline");
         grid.addComponentColumn((item) -> {
            Icon icon;
@@ -147,7 +154,6 @@ public class TaskView extends VerticalLayout {
         grid.asSingleSelect().addValueChangeListener(task -> editTask(task.getValue()));
     }
 
-
     public void editTask(TaskEntity task){
         if (task == null) {
             closeEditor();
@@ -159,6 +165,11 @@ public class TaskView extends VerticalLayout {
 
     public void headerConfig() {
         Header header = new Header();
+    }
+
+    private boolean checkRoleLoggedIn() {
+        return SecurityUtils.getRole().contains(new SimpleGrantedAuthority("ADMIN"))
+                || SecurityUtils.getName().equals(taskForm.task.getOwnerName());
     }
 }
